+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Resource usage check that captures the `top` command output from inside
   the container to the inspect report. CNJR-4618
 - Capture the `/etc/hosts` file from the container and the host. CNJR-7305
+- Capture the container runtime's `network inspect` output. CNJR-7307
 
 ## [0.4.2] - 2025-03-25